+++ src/js/itemset.js
@@ -38,7 +38,7 @@ export { isStaticItemsetFromSecondaryInstance };
 
 export default {
     /**
-     * @param {import('./type-def').UpdatedDataNodes} [updated] - The object containing info on updated data nodes.
+     * @param {UpdatedDataNodes} [updated] - The object containing info on updated data nodes.
      */
     update( updated = {} ) {
         const that = this;