+++ src/widget/date/datepicker-extended.js
@@ -61,7 +61,7 @@ class DatepickerExtended extends Widget {
      * Creates fake date input elements
      *
      * @param {string} format - The date format
-     * @return {import('../../js/src/type-def').jQuery} The jQuery-wrapped fake date input element
+     * @return {jQuery} The jQuery-wrapped fake date input element
      */
     _createFakeDateInput( format ) {
         const $dateI = $( this.element );
@@ -77,7 +77,7 @@ class DatepickerExtended extends Widget {
     /**
      * Copy manual changes that were not detected by bootstrap-datepicker (one without pressing Enter) to original date input field
      *
-     * @param {import('../../js/src/type-def').jQuery} $fakeDateI - Fake date input element
+     * @param {jQuery} $fakeDateI - Fake date input element
      */
     _setChangeHandler( $fakeDateI ) {
         const settings = this.settings;
@@ -119,7 +119,7 @@ class DatepickerExtended extends Widget {
     /**
      * Reset button handler
      *
-     * @param {import('../../js/src/type-def').jQuery} $fakeDateI - Fake date input element
+     * @param {jQuery} $fakeDateI - Fake date input element
      */
     _setResetHandler( $fakeDateI ) {
         $fakeDateI.next( '.btn-reset' ).on( 'click', () => {
@@ -133,7 +133,7 @@ class DatepickerExtended extends Widget {
      * Handler for focus events.
      * These events on the original input are used to check whether to display the 'required' message
      *
-     * @param {import('../../js/src/type-def').jQuery} $fakeDateI - Fake date input element
+     * @param {jQuery} $fakeDateI - Fake date input element
      */
     _setFocusHandler( $fakeDateI ) {
         // Handle focus on original input (goTo functionality)