+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ----------------------
 ##### Fixed
 - Readonly select-minimal widgets are not readonly.
+- Readonly select-minimal widget get erroneously enabled when relevancy changes to positive.
 
 [5.16.0] - 2020-05-28
 ----------------------