+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ##### Fixed
 - Readonly select-minimal widgets are not readonly.
 - Readonly select-minimal widget get erroneously enabled when relevancy changes to positive.
+- Calculation with form control fires input-update event upon load if the underlying model value doesn't change.
 
 [5.16.0] - 2020-05-28
 ----------------------