+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+----------------------
+##### Changed
+- Made fixed English geopicker warning translatable.
+
 [5.16.1] - 2020-06-05
 ----------------------
 ##### Fixed