+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ##### Fixed
 - The "selected" message in the desktop select picker does not switch language immediately.
+- Image-map widget is transparent, not visible but functional, when it is included on a non-first page.
 
 [5.16.1] - 2020-06-05
 ----------------------