+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-[Unreleased]
+[5.16.2] - 2020-06-12
 ----------------------
 ##### Changed
 - Made fixed English geopicker warning translatable.